+++ src/app/layouts/MainLayout.tsx
@@ -1,18 +1,17 @@
-import { Footer } from '@/widgets/footer';
-import { Navbar } from '@/widgets/navbar';
 import { Outlet } from 'react-router-dom';
+import { Navbar, Footer } from '@/widgets';
 import backgroundImage from './assets/background.png';
 
 export const MainLayout = () => {
   return (
-    <div 
+    <div
       className="flex min-h-screen flex-col"
       style={{
         backgroundImage: `url(${backgroundImage})`,
         backgroundSize: 'cover',
         backgroundPosition: 'center',
         backgroundRepeat: 'no-repeat',
-        backgroundAttachment: 'fixed'
+        backgroundAttachment: 'fixed',
       }}
     >
       <Navbar />
