+++ src/pages/home/ui/FilterMatches/FilterMatchesAboutTeam.tsx
@@ -2,13 +2,14 @@ import { CircleQuestionMarkIcon } from 'lucide-react';
 
 import { CustomTooltip } from '@/widgets';
 import { SelectButton } from '@/shared';
+
 import {
-  TEAM_ABILITIES,
-  TEAM_ABILITY_LABELS,
   AGES,
   AGE_LABELS,
   GENDERS,
   GENDER_LABELS,
+  TEAM_ABILITIES,
+  TEAM_ABILITY_LABELS,
   type MatchSearchFilters,
 } from '../../model';
