+++ src/pages/login/ui/LoginPage.tsx
@@ -1,9 +1,10 @@
-import { useState, useCallback } from 'react';
+import { useCallback, useState } from 'react';
 import { Link } from 'react-router-dom';
+
 import { Button } from '@/shared/ui/button';
+import { Checkbox } from '@/shared/ui/checkbox';
 import { Input } from '@/shared/ui/input';
 import { Label } from '@/shared/ui/label';
-import { Checkbox } from '@/shared/ui/checkbox';
 
 export interface LoginFormData {
   email: string;
